Fix-up voice chat for unity 5
Fixed up the voice chat library for unity 5. Changed audio to
GetComponent<AudioSource>() in several locations, added logic to allow
Local debug to work with a single client connected to a server. It is
unclear what the intended functionality of Local debug was at this
point since it is useless if you don’t have a server connected.